+++ sdk/azcore/runtime/policy_http_trace.go
@@ -12,7 +12,6 @@ import (
 	"fmt"
 	"net/http"
 	"net/url"
-	"strings"
 
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/internal/exported"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/internal/shared"
@@ -98,9 +97,6 @@ func (h *httpTracePolicy) Do(req *policy.Request) (resp *http.Response, err erro
 
 // StartSpanOptions contains the optional values for StartSpan.
 type StartSpanOptions struct {
-	// ModuleName is the name of the module that created the span.
-	// If not specified, ModuleName defaults to the azcore module name.
-	ModuleName string
 	// Kind indicates the kind of Span.
 	Kind tracing.SpanKind
 	// Attributes contains key-value pairs of attributes for the span.
@@ -139,9 +135,6 @@ func StartSpan(ctx context.Context, name string, tracer tracing.Tracer, options 
 	if options == nil {
 		options = &StartSpanOptions{}
 	}
-	if options.ModuleName == "" {
-		options.ModuleName = shared.Module
-	}
 	if options.Kind == 0 {
 		options.Kind = tracing.SpanKindInternal
 	}
@@ -154,7 +147,7 @@ func StartSpan(ctx context.Context, name string, tracer tracing.Tracer, options 
 	ctx = context.WithValue(ctx, ctxActiveSpan{}, options.Kind)
 	return ctx, func(err error) {
 		if err != nil {
-			errType := strings.Replace(fmt.Sprintf("%T", err), "*exported.", fmt.Sprintf("*%s.", options.ModuleName), 1)
+			errType := fmt.Sprintf("%T", err)
 			span.SetAttributes(tracing.Attribute{Key: attrErrType, Value: errType})
 			span.SetStatus(tracing.SpanStatusError, fmt.Sprintf("%s:\n%s", errType, err.Error()))
 		}