+++ sdk/azcore/CHANGELOG.md
@@ -15,11 +15,12 @@
 * Added type `Propagator` to the `tracing` package to support context propagation.
 * Added type `Carrier` to the `tracing` package to support context propagation.
 * Added method `NewPropagator()` to type `tracing.Provider` to support context propagation.
-* Added option `ModuleName` to `runtime.StartSpanOptions` to support custom module names in error type.
 
 ### Breaking Changes
 
 ### Bugs Fixed
+
+* Removed hardcoded `azcore` replacement in the error type for `runtime.StartSpan()`.
 
 ### Other Changes